ml-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# PyTorch (CPU-only to save memory)
torch==2.7.0+cpu
torchvision==0.22.0+cpu
numpy==1.26.4
Pillow==10.4.0

# Backend
flask==3.0.3
gunicorn==21.2.0
requests==2.31.0
